Crate str_indices

Count and convert between different indexing schemes on utf8 string slices.

This crate is organized into modules by indexing scheme. Each module contains functions for counting relevant metrics for that scheme as well as functions for converting to/from byte indices.

None of the functions in this crate panic: all inputs have a defined output.

Modules

chars

Index by chars.

lines

Index by lines (all Unicode line breaks).

lines_crlf

Index by lines (carriage return and line feed).

lines_lf

Index by lines (line feed only).

utf16

Index by utf16 code units.